requirements/base.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.0.0
django-redis==5.4.0
redis==5.0.1
django-ratelimit==4.1.0
pytz==2025.2
drf-spectacular==0.28.0
httpx==0.28.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.1.0
websockets==16.0
celery==5.3.0
flower==2.0.1